requirements.txt
streamlit
openai
python-dotenv
google-auth
o365
google-auth
google-auth-oauthlib
google-api-python-client
